REFACTORING: name changed.
st -> flarity

As the fork continues to change drastically, I believe I do not share
some of the suckless philosophy anymore and I do not plan to.

There are many things that this terminal can do and it shouldn't be
limited by the setmind that suckless has.

The credit still remain to them though.